plan: document rendering/halftoning architecture; fix %zu log bug
- Add rendering host note: printer-app can run on Mac/Linux laptop with
  Pi as USB pass-through; laptop renders 5-10x faster than Pi 3B+
- Add rendering/halftoning architecture section with three-path table:
  300 dpi threshold (current), 600 dpi 1-bit + ordered dither, and
  APT Mode 1024 (printer-side halftoning via 8-bit grayscale TIFF)
- Document APT discovery: PCL Mode 1024 with Bits/Sample=8 at 600 dpi
  activates Brother APT; HL-5170DN explicitly listed as APT-capable;
  recommended ≤150 dpi input; potentially best photo quality at ~5s Pi
  render vs 45s for 600 dpi path
- Update Phase 2.1 resolution bullet: 600 dpi requires pgmraw + ordered
  dither (not just -r600), and keepalive must land at the same time
- Rewrite Phase 6 as 6A (APT Mode 1024 investigation) + 6B (HQ1200)
- Fix %zu format specifier in rwriteline diagnostic log: PAPPL logger
  does not handle %zu; cast to (unsigned) and use %u

+++ src/driver.c
@@ -111,8 +111,8 @@ static bool hl5170dn_rwriteline(pappl_job_t *job, pappl_pr_options_t *options,
      * Remove once printing is confirmed working. */
     if (y == 0) {
         papplLogJob(job, PAPPL_LOGLEVEL_INFO,
-            "rwriteline y=0: bytes_per_line=%zu line[0]=0x%02x line[1]=0x%02x",
-            bytes_per_line, line[0], bytes_per_line > 1 ? line[1] : 0u);
+            "rwriteline y=0: bytes_per_line=%u line[0]=0x%02x line[1]=0x%02x",
+            (unsigned)bytes_per_line, line[0], bytes_per_line > 1 ? line[1] : 0u);
     }
 
     encoded_len = packbits_encode(line, bytes_per_line, jd->line_buf);
